Updated language patch code and test
